requirements.txt
-i https://pypi.org/simple
aiohttp==3.9.5; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
asyncio==3.4.3
attrs==23.2.0; python_version >= '3.7'
dacite==1.8.1; python_version >= '3.6'
discord.py==2.4.0; python_version >= '3.8'
dnspython==2.6.1; python_version >= '3.8'
frozenlist==1.4.1; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
motor==3.4.0; python_version >= '3.7'
multidict==6.0.5; python_version >= '3.7'
numpy==2.0.0; python_version >= '3.9'
pygrok==1.0.0
pymongo==4.7.3; python_version >= '3.7'
python-dotenv==1.0.1; python_version >= '3.8'
reactivex==4.0.4; python_version >= '3.7' and python_version < '4.0'
regex==2024.5.15; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
yarl==1.9.4; python_version >= '3.7'
